Split asserting from executing.
