+++ src/lib/Test/Smoke.hs
@@ -1,5 +1,6 @@
 module Test.Smoke
-  ( module Test.Smoke.Bless,
+  ( module Test.Smoke.Assert,
+    module Test.Smoke.Bless,
     module Test.Smoke.Discovery,
     module Test.Smoke.Execution,
     module Test.Smoke.Plan,
@@ -8,6 +9,7 @@ module Test.Smoke
   )
 where
 
+import Test.Smoke.Assert
 import Test.Smoke.Bless
 import Test.Smoke.Discovery
 import Test.Smoke.Execution
