+++ src/lib/Test/Smoke/Types.hs
@@ -2,6 +2,7 @@ module Test.Smoke.Types
   ( module Test.Smoke.Types.Base,
     module Test.Smoke.Types.Errors,
     module Test.Smoke.Types.Executable,
+    module Test.Smoke.Types.ExecutionResult,
     module Test.Smoke.Types.Files,
     module Test.Smoke.Types.Filters,
     module Test.Smoke.Types.Fixtures,
@@ -16,6 +17,7 @@ where
 import Test.Smoke.Types.Base
 import Test.Smoke.Types.Errors
 import Test.Smoke.Types.Executable
+import Test.Smoke.Types.ExecutionResult
 import Test.Smoke.Types.Files
 import Test.Smoke.Types.Filters
 import Test.Smoke.Types.Fixtures
